Enter a game name, and after submitting it, user should be navigated to Search Results Screen
