+++ README.md
@@ -1,15 +1,15 @@
-# C-WIRE-Flutter-challenge
+C-WIRE-Flutter-challenge
+Our objective is to create minimal Flutter app, that is consisting of only 3 screens:
 
-Our objective is to create minimal Flutter app, that is consisting of only 3 screens: 
-
-1) Search / Home 
-2) Search Results List 
-3) Detail Page
-
+Search / Home
+Search Results List
+Detail Page
 Setup:
 
-1) Instal Dart SDK 2.9.0 and Flutter 1.20.1
+Instal Dart SDK 2.9.0 and Flutter 1.20.1
 
-2) Create GitHub account (if you do not have one)
-3) Fork this repository with you GitHub account
-4) Run CWireApplication
+Create GitHub account (if you do not have one)
+
+Fork this repository with you GitHub account
+
+Run CWireApplication
