+++ lib/src/app.dart
@@ -1,3 +1,4 @@
+import 'package:c_wire_app/src/screens/search.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
@@ -16,50 +17,6 @@ class MyApp extends StatelessWidget {
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
       home: SearchScreen(),
-    );
-  }
-}
-
-class SearchScreen extends StatefulWidget {
-  @override
-  _SearchScreenState createState() {
-    return _SearchScreenState();
-  }
-}
-
-class _SearchScreenState extends State<SearchScreen> {
-  TextEditingController _controller;
-
-  void initState() {
-    super.initState();
-    _controller = TextEditingController();
-  }
-
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-          child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-            Padding(
-              padding: const EdgeInsets.all(25.0),
-              child: TextField(
-                controller: _controller,
-                decoration: InputDecoration(
-                    labelText: "Search",
-                    hintText: "Search",
-                    prefixIcon: Icon(Icons.search),
-                    border: OutlineInputBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(25.0)))),
-              ),
-            )
-          ])),
     );
   }
 }
